+++ packages/iov-bcp/src/atomicswaptypes.ts
@@ -2,12 +2,12 @@ import { As } from "type-tagger";
 import { Stream } from "xstream";
 
 import { BcpConnection } from "./connection";
-import { Address, Amount, SwapIdBytes } from "./transactions";
+import { Address, Amount, SwapIdBytes, SwapTimeout } from "./transactions";
 
 export enum SwapState {
   Open = "open",
   Claimed = "claimed",
-  Expired = "expired",
+  Aborted = "aborted",
 }
 
 export interface SwapData {
@@ -21,7 +21,12 @@ export interface SwapData {
    */
   readonly hash: Uint8Array;
   readonly amounts: ReadonlyArray<Amount>;
-  readonly timeout: number; // blockheight where the swap expires (TODO: alternatively support Date?)
+  /**
+   * The first point in time at which the offer is expired.
+   *
+   * Can be represented as a block height or UNIX timestamp.
+   */
+  readonly timeout: SwapTimeout;
   readonly memo?: string;
 }
 
@@ -40,13 +45,13 @@ export interface ClaimedSwap {
   readonly preimage: Preimage;
 }
 
-// ExpiredSwap is an offer that timed out without being claimed
-export interface ExpiredSwap {
-  readonly kind: SwapState.Expired;
+/** A swap offer that has been aborted */
+export interface AbortedSwap {
+  readonly kind: SwapState.Aborted;
   readonly data: SwapData;
 }
 
-export type AtomicSwap = OpenSwap | ClaimedSwap | ExpiredSwap;
+export type AtomicSwap = OpenSwap | ClaimedSwap | AbortedSwap;
 
 export interface AtomicSwapRecipientQuery {
   readonly recipient: Address;